+++ src/main/java/com/coagronet/unidad/dtos/UnidadDTO.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class UnidadDTO {
-    private Integer id;
+    private Long id;
     private String nombre;
     private String descripcion;
-    private Integer estado;
+    private Long estado;
     private Long empresa;
 }
